Authentication

+++ api/routes/catalog.py
@@ -1,7 +1,9 @@
 from typing import List
+from api.authentication.user import get_current_user
+from api.models.user import Submission, User
 
 from beanie import PydanticObjectId
-from fastapi import APIRouter, HTTPException, status
+from fastapi import APIRouter, HTTPException, status, Depends
 
 from api.models.catalog import DatasetMetadataDOC
 
@@ -10,39 +12,57 @@ router = APIRouter()
 
 
 @router.post("/dataset/", response_model=DatasetMetadataDOC, status_code=status.HTTP_201_CREATED)
-async def create_dataset(document: DatasetMetadataDOC):
+async def create_dataset(document: DatasetMetadataDOC, user: User = Depends(get_current_user)):
     await document.insert()
+    submission = document.as_submission()
+    user.submissions.append(submission)
+    await user.save()
     return document
 
 
-@router.get("/dataset/{document_id}", response_model=DatasetMetadataDOC)
-async def get_dataset(document_id: PydanticObjectId):
-    document = await DatasetMetadataDOC.get(document_id)
+@router.get("/dataset/{submission_id}", response_model=DatasetMetadataDOC)
+async def get_dataset(submission_id: PydanticObjectId, user: User = Depends(get_current_user)):
+    submission = user.submission(submission_id)    
+    if submission is None:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Dataset metadata record was not found")
+
+    document = await DatasetMetadataDOC.get(submission.identifier)
     if document is None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Dataset metadata record was not found")
     return document
 
 
 @router.get("/dataset/", response_model=List[DatasetMetadataDOC])
-async def get_datasets():
-    documents = await DatasetMetadataDOC.find_all().to_list()
+async def get_datasets(user: User = Depends(get_current_user)):
+    documents = [await DatasetMetadataDOC.get(submission.identifier) for submission in user.submissions]
     return documents
 
 
-@router.put("/dataset/{document_id}", response_model=DatasetMetadataDOC)
-async def update_dataset(document_id: PydanticObjectId, updated_document: DatasetMetadataDOC):
-    dataset = await DatasetMetadataDOC.get(document_id)
+@router.put("/dataset/{submission_id}", response_model=DatasetMetadataDOC)
+async def update_dataset(submission_id: PydanticObjectId, updated_document: DatasetMetadataDOC, user: User = Depends(get_current_user)):
+    submission = user.submission(submission_id)
+    dataset: DatasetMetadataDOC = await DatasetMetadataDOC.get(submission_id.identifier)
     if dataset is None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Dataset metadata record was not found")
     await dataset.set(updated_document.dict(exclude_unset=True))
-    dataset = await DatasetMetadataDOC.get(document_id)
+    dataset = await DatasetMetadataDOC.get(submission_id)
+    await submission.set(dataset.as_submission().dict(exclude_unset=True))
     return dataset
 
 
-@router.delete("/dataset/{document_id}", response_model=dict)
-async def delete_dataset(document_id: PydanticObjectId):
-    dataset = await DatasetMetadataDOC.get(document_id)
+@router.delete("/dataset/{submission_id}", response_model=dict)
+async def delete_dataset(submission_id: PydanticObjectId, user: User = Depends(get_current_user)):
+    submission = user.submission(submission_id)
+    if submission is None:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Dataset metadata record was not found")
+    dataset = await DatasetMetadataDOC.get(submission.identifier)
     if dataset is None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Dataset metadata record was not found")
+    await submission.delete()
     await dataset.delete()
-    return {"deleted_dataset_id": document_id}
+    return {"deleted_dataset_id": submission_id}
+
+
+@router.get("/submission/", response_model=List[Submission])
+async def get_submissions(user: User = Depends(get_current_user)):
+    return user.submissions
